test(sim): split the owned-class balance harnesses into single-file suites
The level-20 harness ran 788 to 842 seconds as ONE file and the raid
harness 366, so whichever 2-worker CI job drew them was pinned for the
whole duration. Split both by cost cluster into nine files (assertions
moved verbatim; shared diet derivations in tests/helpers/balance_diet.ts,
with the WOC_FULL_BALANCE_SWEEP read kept per-file for the registry pin).

Every split file joins CI_LONG_SUITES and the a/b halves are re-balanced
by the measured per-file durations, so the lane pair and the shard
related legs can finally parallelize the family. Bounds are unchanged
here and re-derive from fresh measured walls in a follow-up change.
