--- tests/owned_class_balance_harness.test.ts
@@ -1,478 +1,0 @@
-import { describe, expect, it } from 'vitest';
-import {
-  averageOwnedClassDpsProbe,
-  averageOwnedHealerProbe,
-  OWNED_CLASS_BALANCE_SCENARIOS,
-  OWNED_CLASS_LEVEL_20_BOSS_SCENARIO,
-  OWNED_CLASS_PBE_LOADOUTS,
-  OWNED_CLASS_PBE_TALENTS,
-  OWNED_DPS_SPECS,
-  type OwnedClassBalanceScenario,
-  runOwnedClassDpsMatrix,
-  runOwnedClassDpsProbe,
-  runOwnedHealerProbe,
-  runWarspiritOfftankProbe,
-} from '../scripts/owned_class_balance_probe';
-import { Sim } from '../src/sim/sim';
-
-// PR-tier diet vs the nightly full sweep (docs/qa-gate.md, "The
-// balance-harness diet"). This suite is a regression tripwire, not a
-// measurement (the authoritative balance instrument is the offline Monte
-// Carlo sweep), so the PR long-sims lane runs a reduced configuration: two of
-// the five fixed seeds, the two band-carrying sustained scenarios of the
-// four-scenario matrix, and 60 s where the full sweep runs 120 s.
-// WOC_FULL_BALANCE_SWEEP=1 (set only by the nightly workflow) restores the
-// full configuration. Every band is pinned to a measurement at ITS OWN
-// configuration via band(full, diet); when a balance change moves a band,
-// re-pin BOTH values from their own printed actuals, and never carry one
-// configuration's value under the other.
-const FULL_SWEEP = process.env.WOC_FULL_BALANCE_SWEEP === '1';
-// Seed VALUES are fixed and never change here; only the count does.
-const BALANCE_SEEDS = FULL_SWEEP
-  ? ([29_930, 29_931, 29_932, 29_933, 29_934] as const)
-  : ([29_930, 29_931] as const);
-const band = (full: number, diet: number): number => (FULL_SWEEP ? full : diet);
-// The 120 s level-20 boss window trips the same coefficient regressions in
-// its first 60 s; the diet halves it. (The raid harness keeps its own 120 s
-// windows: its mana-sustain and cadence assertions are long-fight guards.)
-const BOSS_SCENARIO: OwnedClassBalanceScenario = FULL_SWEEP
-  ? OWNED_CLASS_LEVEL_20_BOSS_SCENARIO
-  : { ...OWNED_CLASS_LEVEL_20_BOSS_SCENARIO, seconds: 60 };
-
-describe('owned-class level 20 balance harness', () => {
-  it('defines the required one-target and three-target burst and sustained scenarios', () => {
-    expect(OWNED_CLASS_BALANCE_SCENARIOS).toEqual([
-      { targets: 1, seconds: 15, window: 'burst' },
-      { targets: 1, seconds: 60, window: 'sustained' },
-      { targets: 3, seconds: 15, window: 'burst' },
-      { targets: 3, seconds: 60, window: 'sustained' },
-    ]);
-  });
-
-  it(
-    'records every requested damage metric for all six owned DPS specs',
-    () => {
-      // Diet: the two sustained scenarios carry every band assertion in this
-      // suite, so the PR tripwire runs them alone; the 15 s burst scenarios
-      // (metadata-only variants of the same rotation loop) ride the nightly
-      // full matrix through runOwnedClassDpsMatrix, which also keeps the
-      // exported matrix entry point itself covered nightly.
-      const metricScenarios = FULL_SWEEP
-        ? OWNED_CLASS_BALANCE_SCENARIOS
-        : [OWNED_CLASS_BALANCE_SCENARIOS[1], OWNED_CLASS_BALANCE_SCENARIOS[3]];
-      const results = FULL_SWEEP
-        ? runOwnedClassDpsMatrix(29_900, 'test-head')
-        : OWNED_DPS_SPECS.flatMap((spec) =>
-            metricScenarios.map((scenario) =>
-              runOwnedClassDpsProbe(spec, scenario, 29_900, 'test-head'),
-            ),
-          );
-      // Literal 8, not OWNED_DPS_SPECS.length: the diet arm builds results
-      // FROM that constant, so a derived expectation would move with any
-      // accidental spec-list shrink instead of catching it (the raid harness
-      // pins its cardinality the same way).
-      expect(results).toHaveLength(8 * metricScenarios.length);
-      expect(new Set(results.map((result) => result.spec))).toEqual(new Set(OWNED_DPS_SPECS));
-      for (const result of results) {
-        expect(result.head).toBe('test-head');
-        expect(result.totalDamage).toBeGreaterThan(0);
-        expect(result.dps).toBe(result.totalDamage / result.scenario.seconds);
-        expect(Object.values(result.damageByTarget)).toHaveLength(result.scenario.targets);
-        expect(Object.values(result.damageByTarget).reduce((sum, value) => sum + value, 0)).toBe(
-          result.totalDamage,
-        );
-        expect(Object.keys(result.damageBySource).length).toBeGreaterThan(0);
-        expect(Object.keys(result.castsByAbility).length).toBeGreaterThan(0);
-        expect(result.buttonsPressed).toBeGreaterThan(0);
-        expect(result.resource.end).toBeGreaterThanOrEqual(0);
-        expect(result.resource.end).toBeLessThanOrEqual(result.resource.max);
-        expect(Object.keys(result.equipment).length).toBeGreaterThan(0);
-        expect(result.equipment).toEqual(OWNED_CLASS_PBE_LOADOUTS[result.spec]);
-        const talents = OWNED_CLASS_PBE_TALENTS[result.spec];
-        if (talents) expect(result.talents).toEqual(talents);
-        expect(result.dualWielding).toBe(result.spec === 'warspirit');
-      }
-      const vespersArea = results.find(
-        (result) =>
-          result.spec === 'vespers' &&
-          result.scenario.targets === 3 &&
-          result.scenario.seconds === 60,
-      );
-      expect(vespersArea?.damageByTarget.target_2).toBeGreaterThan(0);
-      expect(vespersArea?.damageByTarget.target_3).toBeGreaterThan(0);
-      const thundercallArea = results.find(
-        (result) =>
-          result.spec === 'thundercall' &&
-          result.scenario.targets === 3 &&
-          result.scenario.seconds === 60,
-      );
-      expect(thundercallArea?.damageByTarget.target_2).toBeGreaterThan(0);
-      expect(thundercallArea?.damageByTarget.target_3).toBeGreaterThan(0);
-      expect(thundercallArea?.castsByAbility.Skybranch).toBeGreaterThan(0);
-      const moongroveArea = results.find(
-        (result) =>
-          result.spec === 'moongrove' &&
-          result.scenario.targets === 3 &&
-          result.scenario.seconds === 60,
-      );
-      expect(moongroveArea?.damageByTarget.target_2).toBeGreaterThan(0);
-      expect(moongroveArea?.damageByTarget.target_3).toBeGreaterThan(0);
-      // The payoff is a CHOICE (Moonsurge or Sunwake) since Moongrove v3, so a
-      // short window may legitimately never pick the sun; both-arm coverage is
-      // pinned by the druid_engines parity scenario, which presses each.
-      expect(
-        (moongroveArea?.castsByAbility.Moonsurge ?? 0) +
-          (moongroveArea?.castsByAbility.Sunwake ?? 0),
-      ).toBeGreaterThan(0);
-      const wildfangSustained = results.find(
-        (result) =>
-          result.spec === 'wildfang' &&
-          result.scenario.targets === 1 &&
-          result.scenario.seconds === 60,
-      );
-      expect(wildfangSustained?.castsByAbility.Redharvest).toBeGreaterThan(0);
-      // The Stampede cooldown tripwire rides the sustained window at PR time
-      // (a 60 s window contains the opener the burst window pinned); the burst
-      // window's own copy runs on the nightly full matrix.
-      const packlordSustained = results.find(
-        (result) =>
-          result.spec === 'packlord' &&
-          result.scenario.targets === 1 &&
-          result.scenario.seconds === 60,
-      );
-      expect(packlordSustained?.castsByAbility.Stampede).toBeGreaterThan(0);
-      expect(packlordSustained?.damageBySource.Stampede).toBeGreaterThan(0);
-      if (FULL_SWEEP) {
-        const packlordBurst = results.find(
-          (result) =>
-            result.spec === 'packlord' &&
-            result.scenario.targets === 1 &&
-            result.scenario.seconds === 15,
-        );
-        expect(packlordBurst?.castsByAbility.Stampede).toBeGreaterThan(0);
-        expect(packlordBurst?.damageBySource.Stampede).toBeGreaterThan(0);
-      }
-      // OWNED_DPS_SPECS grew 6 -> 8 with the druid overhaul (moongrove/wildfang).
-      // Diet budget: ~67s measured local; 300s keeps the ~2.5x fast-runner
-      // margin plus lane headroom.
-    },
-    FULL_SWEEP ? 480_000 : 300_000,
-  );
-
-  it('is deterministic at the same fixed seed and fixture', () => {
-    const scenario = OWNED_CLASS_BALANCE_SCENARIOS[3];
-    expect(runOwnedClassDpsProbe('fieldcraft', scenario, 29_901)).toEqual(
-      runOwnedClassDpsProbe('fieldcraft', scenario, 29_901),
-    );
-  }, 120_000);
-
-  it('pins a Fieldcraft sustained-damage ceiling against the ranged Hunter specs and pays Bloodhook', () => {
-    const scenario = OWNED_CLASS_BALANCE_SCENARIOS[1];
-    const coldsight = runOwnedClassDpsProbe('coldsight', scenario, 29_902);
-    const fieldcraft = runOwnedClassDpsProbe('fieldcraft', scenario, 29_902);
-    const woundDamage = fieldcraft.damageBySource['Bloodhook Wound'] ?? 0;
-
-    // Band widened for the stacked v0.29 rogue redesign (#2328): its shared
-    // combat changes shift this pair a few percent; re-author when it lands.
-    // Ceiling only, deliberately: there is no matching floor here pending the
-    // Hunter kit debt, so a real downside swing is allowed to pass.
-    expect(fieldcraft.dps).toBeLessThanOrEqual(coldsight.dps * 1.25);
-    expect(woundDamage / fieldcraft.totalDamage).toBeGreaterThanOrEqual(0.05);
-  }, 120_000);
-
-  it('keeps Vespers sustained damage in the DPS caster band', () => {
-    const scenario = OWNED_CLASS_BALANCE_SCENARIOS[1];
-    const thundercall = runOwnedClassDpsProbe('thundercall', scenario, 29_903);
-    const vespers = runOwnedClassDpsProbe('vespers', scenario, 29_903);
-
-    expect(vespers.dps).toBeGreaterThanOrEqual(thundercall.dps * 0.9);
-    // Band widened for the stacked v0.29 rogue redesign (#2328): its shared
-    // combat changes shift this pair a few percent; re-author when it lands.
-    expect(vespers.dps).toBeLessThanOrEqual(thundercall.dps * 1.2);
-  }, 120_000);
-
-  it(
-    'keeps the fixed Shaman and Vespers builds inside their sustained role bands',
-    () => {
-      const single = OWNED_CLASS_BALANCE_SCENARIOS[1];
-      const area = OWNED_CLASS_BALANCE_SCENARIOS[3];
-      const thundercall = averageOwnedClassDpsProbe('thundercall', single, BALANCE_SEEDS);
-      const warspiritSingle = averageOwnedClassDpsProbe('warspirit', single, BALANCE_SEEDS);
-      const warspiritArea = averageOwnedClassDpsProbe('warspirit', area, BALANCE_SEEDS);
-      const vespersSingle = averageOwnedClassDpsProbe('vespers', single, BALANCE_SEEDS);
-      const vespersArea = averageOwnedClassDpsProbe('vespers', area, BALANCE_SEEDS);
-      const warspiritBoss = averageOwnedClassDpsProbe('warspirit', BOSS_SCENARIO, BALANCE_SEEDS);
-      const vespersBoss = averageOwnedClassDpsProbe('vespers', BOSS_SCENARIO, BALANCE_SEEDS);
-
-      // Floor lowered for the v0.36 composition (Vespers re-band landed Shadow
-      // at ~214; Elemental is a below-band kit item tracked separately);
-      // flagged for owner review. Lane-diet re-measure: full actual 0.9612 (5
-      // seeds), diet actual 0.9663 (2 seeds); same relative margin keeps the
-      // 0.83 floor and puts the diet ceiling at 1.11.
-      expect(thundercall.dps).toBeGreaterThanOrEqual(vespersSingle.dps * 0.83);
-      expect(thundercall.dps).toBeLessThanOrEqual(vespersSingle.dps * band(1.1, 1.11));
-      // Warspirit area/single: full actual 1.1494, diet actual 1.0944 (the two
-      // retained seeds roll the single-target run high), so the diet band is
-      // 1.04 to 1.14 at the same relative margins.
-      expect(warspiritArea.dps / warspiritSingle.dps).toBeGreaterThanOrEqual(band(1.1, 1.04));
-      expect(warspiritArea.dps / warspiritSingle.dps).toBeLessThanOrEqual(band(1.2, 1.14));
-      // Vespers area/single: full actual 1.4041, diet actual 1.4475; the diet
-      // floor rises to 1.29 with the same relative margin.
-      expect(vespersArea.dps / vespersSingle.dps).toBeGreaterThanOrEqual(band(1.25, 1.29));
-      // 2026-08-09 120s band round: the Warspirit raise (stormstrike row plus
-      // the baseline AP arm, ridden on apPct after review) and the Vespers trim
-      // moved this pair to a measured 1.1539 (warspirit 204.5 / vespers 177.2),
-      // so the 0.93 floor is green again with real margin. Lane-diet
-      // re-measure: full actual 1.1539 (5 seeds, 120 s boss), diet actual
-      // 1.1775 (2 seeds, 60 s boss); same relative margins give 0.95 / 1.22.
-      expect(warspiritBoss.dps / vespersBoss.dps).toBeGreaterThanOrEqual(band(0.93, 0.95));
-      // Full-sweep ceiling kept at 1.2 (measured 1.1539 that round, was 1.18
-      // on the combined tree pre-round). Re-author both sides of this pair
-      // when the owned-class stack integrates.
-      expect(warspiritBoss.dps / vespersBoss.dps).toBeLessThanOrEqual(band(1.2, 1.22));
-      // Full sweep: the grown owned-class matrix ran ~180s under shard load and
-      // roughly doubled in the shared lane (run 31288946173 killed it at 240s).
-      // Diet: two seeds and the 60 s boss window cut the simulated time 3.2x.
-    },
-    FULL_SWEEP ? 900_000 : 300_000,
-  );
-
-  it(
-    'keeps the Druid damage arms sane on the fixed low-SP probe',
-    () => {
-      // IMPORTANT: this fixed PBE loadout is a level-20 caster PROXY (spell power
-      // ~105). Balance's damage was re-seated onto spell-power coefficients, so on
-      // the real searched best-in-slot of the endgame tree (spell power ~150) it
-      // scales to the ~200 DPS anchor measured by the Nythraxis montecarlo, and the
-      // coefficients are calibrated to that. On this low-SP proxy it reads ~155.
-      // The melee Wildfang cat (agility) is NOT under-geared here, so the two arms
-      // are not directly comparable on the proxy: Balance/Feral parity at real BiS
-      // is owned by the montecarlo, not this probe. These bands only guard the
-      // proxy against gross regression. Diet window 60 s (a gross proxy
-      // regression shows in the first minute); the nightly sweep keeps 120 s.
-      const scenario: OwnedClassBalanceScenario = {
-        targets: 1,
-        seconds: FULL_SWEEP ? 120 : 60,
-        window: 'raid',
-      };
-      const moongrove = runOwnedClassDpsProbe('moongrove', scenario, 29_904);
-      const wildfang = runOwnedClassDpsProbe('wildfang', scenario, 29_904);
-
-      // Lane-diet re-measure: moongrove full actual 154.00 (120 s), diet
-      // actual 152.33 (60 s); wildfang 176.08 / 179.05. Diet bands re-derived
-      // at the same relative margins: 137 to 178 and 168 to 208.
-      expect(moongrove.dps).toBeGreaterThanOrEqual(band(138, 137));
-      expect(moongrove.dps).toBeLessThanOrEqual(band(180, 178));
-      expect(wildfang.dps).toBeGreaterThanOrEqual(band(165, 168));
-      expect(wildfang.dps).toBeLessThanOrEqual(band(205, 208));
-    },
-    FULL_SWEEP ? 180_000 : 90_000,
-  );
-
-  it(
-    'keeps Moongrove naked damage within 15% of the naked peer band',
-    () => {
-      // The v0.29 Balance rebalance shifted Moongrove's power off flat base
-      // numbers onto spell-power coefficients, so an un-geared caster scales down
-      // to the pack instead of towering over it the way the old flat numbers did
-      // (pre-rebalance naked Moongrove was the single highest naked spec, ~+50%).
-      // Measured with no gear against the boss-flag dummy, the gear-by-measurement
-      // axis the balance guide uses, averaged to shed per-seed rotation noise.
-      const scenario = {
-        targets: 1,
-        seconds: 60,
-        window: 'raid',
-        targetLevel: 20,
-        targetTemplateId: 'nythraxis_scourge_of_thornpeak',
-      } as const;
-      // Seed 29_932 degenerately stalls the Moongrove rotation on this bench; the
-      // other three are stable (the diet keeps the first two of them).
-      const seeds = FULL_SWEEP ? ([29_904, 29_930, 29_931] as const) : ([29_904, 29_930] as const);
-      const nakedAvg = (spec: Parameters<typeof runOwnedClassDpsProbe>[0]) =>
-        seeds.reduce(
-          (sum, seed) =>
-            sum + runOwnedClassDpsProbe(spec, scenario, seed, 'naked', undefined, 'naked').dps,
-          0,
-        ) / seeds.length;
-      const moongrove = nakedAvg('moongrove');
-      // Two un-geared peers: a ranged pet spec and a caster, the band Moongrove
-      // must sit inside rather than above. Lane-diet re-measure: full actual
-      // ratio 0.9057 (3 seeds), diet actual 0.8754 (2 seeds); the diet band is
-      // 0.82 to 1.11 at the same relative margins.
-      const peerBand = (nakedAvg('packlord') + nakedAvg('vespers')) / 2;
-      expect(moongrove / peerBand).toBeLessThanOrEqual(band(1.15, 1.11));
-      expect(moongrove / peerBand).toBeGreaterThanOrEqual(band(0.85, 0.82));
-    },
-    FULL_SWEEP ? 240_000 : 180_000,
-  );
-
-  it.each(['spiritmend', 'doctrine', 'benison', 'groveheart'] as const)(
-    'records the fixed one-ally and three-ally %s healing profiles',
-    (spec) => {
-      for (const allies of [1, 3] as const) {
-        const result = runOwnedHealerProbe(spec, allies, 29_910, 'test-head');
-        expect(result.head).toBe('test-head');
-        expect(result.effectiveHealing).toBeGreaterThan(0);
-        expect(result.hps).toBe(result.effectiveHealing / result.seconds);
-        expect(result.overhealing).toBeGreaterThanOrEqual(0);
-        expect(result.overhealPct).toBeGreaterThanOrEqual(0);
-        expect(result.overhealPct).toBeLessThanOrEqual(1);
-        expect(result.emergencyRecoverySeconds).not.toBeNull();
-        expect(result.resource.end).toBeGreaterThanOrEqual(0);
-        expect(Object.keys(result.castsByAbility).length).toBeGreaterThan(0);
-        expect(Object.keys(result.equipment).length).toBeGreaterThan(0);
-        expect(result.talents).toEqual(OWNED_CLASS_PBE_TALENTS[spec]);
-      }
-    },
-    30_000,
-  );
-
-  it(
-    'keeps each healer build inside its seed-averaged role and mana contract',
-    () => {
-      const spiritmendSingle = averageOwnedHealerProbe('spiritmend', 1, BALANCE_SEEDS);
-      const spiritmendGroup = averageOwnedHealerProbe('spiritmend', 3, BALANCE_SEEDS);
-      const doctrineSingle = averageOwnedHealerProbe('doctrine', 1, BALANCE_SEEDS);
-      const doctrineGroup = averageOwnedHealerProbe('doctrine', 3, BALANCE_SEEDS);
-      const benisonSingle = averageOwnedHealerProbe('benison', 1, BALANCE_SEEDS);
-      const benisonGroup = averageOwnedHealerProbe('benison', 3, BALANCE_SEEDS);
-
-      // Lane-diet re-measure (full actuals at 5 seeds / diet at 2): the healer
-      // probes are nearly seed-stable, so most same-relative-margin re-pins
-      // land back on the full values at the diet's granularity: benison
-      // recovery 4.75 in both, benisonGroup/spiritmendGroup hps ratio 1.0874 /
-      // 1.0837 (floor stays 0.8), benison resourceEnd 924.0+982.2 / 924.0+986.5
-      // (floors stay 250), spiritmendGroup resourceEnd 2234.2 / 2249.5 (floor
-      // stays 1_200), doctrineSingle hps+dps 155.31 / 154.48 (floor stays 140),
-      // doctrineGroup resourceEnd 719.6 / 727.5 (floor stays 150). The one
-      // mover: doctrineGroup hps+dps+absorbed/60 measured 168.06 full / 182.28
-      // diet, so the diet floor is 130.
-      expect(benisonGroup.emergencyRecoverySeconds).toBeLessThan(
-        spiritmendGroup.emergencyRecoverySeconds,
-      );
-      expect(benisonGroup.hps).toBeGreaterThanOrEqual(spiritmendGroup.hps * 0.8);
-      expect(benisonSingle.resourceEnd).toBeGreaterThanOrEqual(250);
-      expect(benisonGroup.resourceEnd).toBeGreaterThanOrEqual(250);
-      expect(spiritmendGroup.resourceEnd).toBeGreaterThanOrEqual(1_200);
-      expect(doctrineSingle.hps + doctrineSingle.dps).toBeGreaterThanOrEqual(140);
-      expect(
-        doctrineGroup.hps + doctrineGroup.dps + doctrineGroup.absorbedDamage / 60,
-      ).toBeGreaterThanOrEqual(band(120, 130));
-      expect(doctrineGroup.resourceEnd).toBeGreaterThanOrEqual(150);
-      expect(spiritmendSingle.hps).toBeGreaterThan(0);
-      // Same owned-class matrix growth as the DPS metric test above, same
-      // long-sims lane contention doubling; the diet runs two of the five
-      // seeds (~50s measured local).
-    },
-    FULL_SWEEP ? 720_000 : 240_000,
-  );
-
-  it('runs Priest healer pressure through shields and Seraphic Vigil', () => {
-    const doctrine = runOwnedHealerProbe('doctrine', 3, 29_912);
-    const benison = runOwnedHealerProbe('benison', 3, 29_912);
-
-    expect(doctrine.absorbedDamage).toBeGreaterThan(0);
-    // The pressure run must still WEAVE the Vigil into the rotation; whether
-    // it fires is the party's health, asserted deterministically below (a
-    // live benison healer keeps the probe party above the 35% trigger for
-    // whole runs, so a triggered-heal assertion here was flaky-by-design).
-    expect(benison.castsByAbility['Seraphic Vigil'] ?? 0).toBeGreaterThan(0);
-
-    // The trigger contract, exercised directly: ward an ally, drop them below
-    // the 35% threshold with one hit, and the consumed Vigil pays its heal as
-    // an attributable Seraphic Vigil healing event.
-    const sim = new Sim({ seed: 29_912, playerClass: 'priest', autoEquip: true }) as Sim & {
-      drainEvents(): { type: string; ability?: string; amount?: number }[];
-      ctx: {
-        dealDamage(
-          source: unknown,
-          target: unknown,
-          amount: number,
-          direct: boolean,
-          school: string,
-          ability: string,
-          outcome: string,
-        ): void;
-      };
-    };
-    sim.setPlayerLevel(20);
-    expect(sim.setSpec('holy')).toBe(true);
-    const priest = sim.player;
-    priest.resource = priest.maxResource;
-    sim.targetEntity(priest.id);
-    sim.castAbility('seraphic_vigil');
-    sim.tick();
-    expect(priest.auras.some((aura) => aura.id === 'seraphic_vigil')).toBe(true);
-    priest.hp = Math.floor(priest.maxHp * 0.4);
-    sim.drainEvents();
-    sim.ctx.dealDamage(
-      null,
-      priest,
-      Math.floor(priest.maxHp * 0.1),
-      false,
-      'physical',
-      'Vigil Probe',
-      'hit',
-    );
-    const vigilHeal = sim
-      .drainEvents()
-      .filter((event): event is Extract<typeof event, { type: 'heal2' }> => event.type === 'heal2')
-      .find((event) => event.ability === 'Seraphic Vigil');
-    expect(vigilHeal?.amount ?? 0).toBeGreaterThan(0);
-    expect(priest.auras.some((aura) => aura.id === 'seraphic_vigil')).toBe(false);
-  }, 120_000);
-
-  it('counts Groveheart heal-over-time ticks in the effective-healing profile', () => {
-    const groveheart = runOwnedHealerProbe('groveheart', 3, 29_913);
-
-    expect(groveheart.healingBySource.Wildbloom).toBeGreaterThan(0);
-    expect(groveheart.hps).toBeGreaterThan(0);
-  });
-
-  it('holds the Groveheart interim healer contract on both profiles', () => {
-    // Single target: inside the peer envelope at the shared seed.
-    const singlePeers = (['spiritmend', 'doctrine', 'benison'] as const).map(
-      (spec) => runOwnedHealerProbe(spec, 1, 29_914).hps,
-    );
-    const single = runOwnedHealerProbe('groveheart', 1, 29_914).hps;
-    expect(single).toBeGreaterThanOrEqual(Math.min(...singlePeers));
-    expect(single).toBeLessThanOrEqual(Math.max(...singlePeers) * 1.15);
-
-    // Group profile: INTERIM floor, not the envelope. The v0.31 healer
-    // retunes lifted every peer's three-ally throughput while Groveheart
-    // still carries its v0.29 values, and under the heavier pressure the
-    // garden never plants (pure triage). Closing that gap is the flagged
-    // PBE values pass for the druid stack; this floor only guards against
-    // regressions below the measured interim state.
-    const groupPeers = (['spiritmend', 'doctrine', 'benison'] as const).map(
-      (spec) => runOwnedHealerProbe(spec, 3, 29_914).hps,
-    );
-    const group = runOwnedHealerProbe('groveheart', 3, 29_914).hps;
-    expect(group).toBeGreaterThanOrEqual(Math.min(...groupPeers) * 0.45);
-    expect(group).toBeLessThanOrEqual(Math.max(...groupPeers) * 1.15);
-
-    // Absolute floors so the whole band cannot sink together unnoticed: the
-    // agility-loadout regression measured 65.0 and 26.2 here.
-    expect(single).toBeGreaterThanOrEqual(80);
-    expect(group).toBeGreaterThanOrEqual(40);
-  }, 300_000);
-
-  it('records Warspirit mitigation, threat, forced-target uptime, and exit behavior', () => {
-    const result = runWarspiritOfftankProbe(29_920, 'test-head');
-    expect(result.head).toBe('test-head');
-    expect(result.stoneboundIncomingDamage).toBeLessThan(result.galeheartIncomingDamage);
-    expect(result.stoneboundMitigationPct).toBeGreaterThan(0);
-    expect(result.stoneboundThreatFrom100Damage).toBeGreaterThanOrEqual(200);
-    expect(result.forcedTargetUptimeSeconds).toBeGreaterThanOrEqual(3);
-    expect(result.forcedTargetUptimeSeconds).toBeLessThanOrEqual(3.1);
-    expect(result.secondsToLoseThreatAfterLeaving).toBeGreaterThan(0);
-    expect(result.secondsToLoseThreatAfterLeaving).toBeLessThanOrEqual(60);
-  });
-
-  it('keeps role probes deterministic at the same fixed seed', () => {
-    expect(runOwnedHealerProbe('spiritmend', 3, 29_911)).toEqual(
-      runOwnedHealerProbe('spiritmend', 3, 29_911),
-    );
-    expect(runWarspiritOfftankProbe(29_921)).toEqual(runWarspiritOfftankProbe(29_921));
-  }, 120_000);
-});
